master
- Guard seekTo against null contentWindow before iframe finishes loading
- Validate source URLs with new URL() constructor before rendering links
- Show error state instead of blank skeleton when video/channel path is empty
- Use float for video duration to match DB REAL schema and preserve precision

+++ backend/services/transcript_service.py
@@ -86,7 +86,7 @@ def extract_transcript(youtube_url: str, max_duration_seconds: int = 600) -> Tra
     except Exception as e:
         raise TranscriptError(f"Could not fetch video info: {str(e)}") from e
 
-    duration = int(info.get("duration") or 0)
+    duration = float(info.get("duration") or 0)
     if not duration:
         raise TranscriptError("Could not determine video duration.")
     if duration > max_duration_seconds:
